pyproject.toml
[project]
name = "ai-assistant-rag-langchain"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-huggingface>=1.2.1",
    "langgraph>=0.2.0",
    "modelscope>=1.35.3",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "sentencepiece>=0.2.1",
    "torch>=2.11.0",
    "transformers>=5.5.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
